Added StonesInPlay
