+++ code/src/sneer/bricks/snapps/games/go/gui/graphics/HoverStone.java
@@ -3,7 +3,6 @@ package sneer.bricks.snapps.games.go.gui.graphics;
 import java.awt.Color;
 import java.awt.Graphics2D;
 import java.awt.RenderingHints;
-import java.awt.geom.Ellipse2D;
 
 import sneer.bricks.snapps.games.go.GoBoard;
 import sneer.bricks.snapps.games.go.GoBoard.StoneColor;
@@ -11,6 +10,12 @@ import sneer.bricks.snapps.games.go.gui.GoBoardPanel;
 
 public class HoverStone{
 
+	private StonePainter stonePainter;
+
+	public HoverStone(final StonePainter stonePainter) {
+		this.stonePainter = stonePainter;
+	}
+	
 	public void draw(final Graphics2D graphics, final GoBoard _board, final int _hoverX, final int _hoverY,final int _scrollX, final int _scrollY){
 		if (!_board.canPlayStone(unscrollX(_hoverX,_scrollX), unscrollY(_hoverY, _scrollY))) return;
 
@@ -18,7 +23,7 @@ public class HoverStone{
 		else graphics.setColor(new Color(255, 255, 255, 90));
 			
 		graphics.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-		paintStoneOnCoordinates(graphics, toCoordinateSmall(_hoverX), toCoordinateSmall(_hoverY), false);
+		stonePainter.paintStoneOnCoordinates(graphics, toCoordinateSmall(_hoverX), toCoordinateSmall(_hoverY), false);
 	}
 
 	private float toCoordinateSmall(int position) {
@@ -33,22 +38,5 @@ public class HoverStone{
 		return (GoBoardPanel.BOARD_SIZE + y - _scrollY) % GoBoardPanel.BOARD_SIZE; 
 	}
 	
-	private void paintStoneOnCoordinates(Graphics2D graphics, float x, float y, boolean dead) {
-		float d = GoBoardPanel.STONE_DIAMETER;
-		if (dead) d*=0.6;
-		
-		graphics.fill(new Ellipse2D.Float(x - (d / 2), y - (d / 2), d, d));
-		//wrapping
-		int buffersize=(int)(GoBoardPanel.BOARD_IMAGE_SIZE+GoBoardPanel.CELL_SIZE);
-		
-		if (x==0) graphics.fill(new Ellipse2D.Float(buffersize - (d / 2), y - (d / 2), d, d));
-		if (y==0) graphics.fill(new Ellipse2D.Float(x - (d / 2), buffersize - (d / 2), d, d));
-		if (x==buffersize) graphics.fill(new Ellipse2D.Float(- (d / 2), y - (d / 2), d, d)); 
-		if (y==buffersize) graphics.fill(new Ellipse2D.Float(x - (d / 2), - (d / 2), d, d)); 
-		
-		if (x==0 & y==0) graphics.fill(new Ellipse2D.Float(buffersize - (d / 2), buffersize  - (d / 2), d, d));
-		if (x==buffersize & y==0) graphics.fill(new Ellipse2D.Float(- (d / 2), buffersize  - (d / 2), d, d));
-		if (x==buffersize & y==buffersize) graphics.fill(new Ellipse2D.Float(- (d / 2), - (d / 2), d, d));
-		if (x==0 & y==buffersize) graphics.fill(new Ellipse2D.Float(buffersize- (d / 2), - (d / 2), d, d));
-	}
+	
 }
